+++ spec/rubocop/cop/lint/shadowed_exception_spec.rb
@@ -122,6 +122,37 @@ describe RuboCop::Cop::Lint::ShadowedException do
 
       expect(cop.offenses).to be_empty
     end
+
+    it 'accepts rescuing nil' do
+      inspect_source(cop, ['begin',
+                           '  a',
+                           'rescue nil',
+                           '  b',
+                           'end'])
+
+      expect(cop.offenses).to be_empty
+    end
+
+    it 'accepts rescuing nil and another exception' do
+      inspect_source(cop, ['begin',
+                           '  a',
+                           'rescue nil, Exception',
+                           '  b',
+                           'end'])
+
+      expect(cop.offenses).to be_empty
+    end
+
+    it 'registers an offense when rescuing nil multiple exceptions of ' \
+       'different levels' do
+      inspect_source(cop, ['begin',
+                           '  a',
+                           'rescue nil, StandardError, Exception',
+                           '  b',
+                           'end'])
+
+      expect(cop.messages).to eq(['Do not shadow rescued Exceptions'])
+    end
   end
 
   context 'multiple rescues' do
@@ -256,6 +287,30 @@ describe RuboCop::Cop::Lint::ShadowedException do
 
         expect(cop.offenses).to be_empty
       end
+    end
+
+    it 'accepts rescuing nil before another exception' do
+      inspect_source(cop, ['begin',
+                           '  a',
+                           'rescue nil',
+                           '  b',
+                           'rescue',
+                           '  c',
+                           'end'])
+
+      expect(cop.offenses).to be_empty
+    end
+
+    it 'accepts rescuing nil after another exception' do
+      inspect_source(cop, ['begin',
+                           '  a',
+                           'rescue',
+                           '  b',
+                           'rescue nil',
+                           '  c',
+                           'end'])
+
+      expect(cop.offenses).to be_empty
     end
   end
 end